SaucoBot/requirements.txt
discord.py
requests~=2.31.0
pytz~=2024.1
discord~=2.3.2
openai~=1.12.0
python-dotenv
pandas
psycopg2-binary
